devtools/server/tests/mochitest/test_framerate_05.html

<!DOCTYPE HTML>
<html>
<!--
Bug 1034648 - Tests whether a framerate recording can be cancelled.
-->
<head>
  <meta charset="utf-8">
  <title>Framerate actor test</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script>
"use strict";

window.onload = function() {
  const {require} = ChromeUtils.import("resource://devtools/shared/Loader.jsm", {});
  const {DebuggerClient} = require("devtools/shared/client/debugger-client");
  const {DebuggerServer} = require("devtools/server/main");
  const Services = require("Services");

  // Always log packets when running tests.
  Services.prefs.setBoolPref("devtools.debugger.log", true);
  SimpleTest.registerCleanupFunction(function() {
    Services.prefs.clearUserPref("devtools.debugger.log");
  });

  SimpleTest.waitForExplicitFinish();

  const {FramerateFront} = require("devtools/shared/fronts/framerate");

  DebuggerServer.init();
  DebuggerServer.registerAllActors();

  const client = new DebuggerClient(DebuggerServer.connectPipe());
  client.connect().then(function onConnect() {
    client.listTabs().then(function onListTabs(response) {
      const form = response.tabs[response.selected];
      const front = FramerateFront(client, form);

      front.startRecording().then(() => {
        /* eslint-disable max-nested-callbacks */
        window.setTimeout(() => {
          front.cancelRecording().then(() => {
            window.setTimeout(() => {
              front.getPendingTicks().then(rawTicks => {
                ok(rawTicks,
                  "The returned pending ticks should be empty (1).");
                is(rawTicks.length, 0,
                  "The returned pending ticks should be empty (2).");

                front.stopRecording().then(rawData => {
                  ok(rawData,
                    "The returned raw data should be an empty array (1).");
                  is(rawData.length, 0,
                    "The returned raw data should be an empty array (2).");

                  client.close().then(() => {
                    DebuggerServer.destroy();
                    SimpleTest.finish();
                  });
                });
              });
            }, 1000);
          });
        }, 1000);
        /* eslint-enable max-nested-callbacks */
      });
    });
  });
};
</script>
</pre>
</body>
</html>
